requirements.txt
streamlit
google-generativeai>=0.3.0
pdfplumber
scikit-learn
requests
aiohttp
faiss-cpu
numpy
pandas
sentence-transformers
python-dotenv
arxiv
semanticscholar